v 1.8.4
New features:

1.
Welcome Screen with the ability to choose useful templates for a quick start in working with the application.

2.
Five updated templates that will help you be more focused and creative.

3.
Editable list of the recent documents.

4.
Auto-saving document mode. A long-awaited basic feature.

5.
Function: "Open multiple documents with 1 click".

6.
Function: "Open documents on separate desktops in full screen mode". However, it's only available for MacOS at the moment. I'm not sure how to implement it for Windows yet.
Watch a short video https://youtube.com/shorts/iK6NGHoXNYU

7.
Now ProtoText is compatible with the ChatGPT 3.5-turbo model and therefore with GPT-4. If you have access to this version of the API, let me know and let's test it.

8.
The parameter "Maximum number of tokens" has been replaced with "Limit Tokens". In simpler terms, enter 0 in this field to not worry about tokens and use the maximum available in the selected model.

Minor changes:

- Notifications of requests errors to ChatGPT.
- The application loading process is now equipped with an indicator.
- Bug fixes.