v 1.9.0 Hello friends! We have many updates.
*1. Advanced integrations.*
Now you can have more flexible control over AI/APIs integrations in Prototext and add new ones according to your preference.

The new feature is called "Actions".

At the moment, there are 7 templates available for actions.

With these 4, you are already familiar:
- Text generation by OpenAI.
- Text translation by OpenAI.
- Text translation by Microsoft.
- Speech synthesis by ElevenLabs.

This is a new addition from the same series for audio processing:
- Speech recognition by OpenAI.

And these two are the freshest and most interesting:
- Custom API endpoint.
- Custom script.

Custom API actions open the possibility to connect any external API. For example, for working with open-source neural networks through Google Colab, HuggingFace 🤗, or software running locally on your computer. Currently, the available actions include integrating Conqui TTS 🐸 (free private software for speech synthesis) and LLaMA-cpp 🦙 (also free and private software for working with large language models). Installation tutorials are included in the app.

Custom API actions implement the transfer and reception of text, images, audio, video, and any other data formats. Yes, now you can send a request even A1111 with StableDiffusion to your Google Colab notebook to generate an image and save it in your PTXT document or request weather forecasts or exchange rates and format the responses into new cards.
Custom scripts are an action for your JavaScript code to work with card content. In the demo action, you will find a simple snippet to add the current date to the target card. This way, you can transform any content in any desired way.

The number of actions is not limited. You can create duplicates of them with fine-tuned settings. For example, to synthesize speech with different voices, generate and translate texts with narrow specialized context.

The appearance of actions is customizable. It can be a button with text or an emoji icon according to your preference.
Move your cursor over the icons with question marks (?) to read tooltips for all important action settings.

*2. Audio recording.*
Enter "@" inside an empty card and select the "Audio recording" option. ProtoText will ask for microphone permission. The recorded audio in mp3 format can be transcribed or used as audio notes. Secretly, I'll tell you that I'm currently developing another big feature - ProtoText can become a tool for recording podcasts and even a little video editor. But more on that later ;)

*3. Speech recognition.*
I have long wanted to integrate Whisper. Well, now it's on board. Moreover, with the help of the custom action Whisper, you can use it locally and for free on your computer or in Google Colab.

*4. New in interaction with actions (AI actions).*
If you have ever used ProtoText and the famous "AI" button, then you know that it is a request to ChatGPT. It always takes 2 clicks, and the chat response is always inside 1 card with a rainbow stripe on the left side. Requests can only be made sequentially.

What now?

- 1 click or 2 to create a prompt - at your discretion in the action settings.

- The chat response can be automatically divided into separate cards by paragraphs, also look for this in the settings.

- Parallel requests 🥳 hopefully without bugs, now you can make any number of requests at the same time. For example, asking ChatGPT and generating images.

- Error reports without pop-up windows, just inside the answer card.

- If you didn't expect an AI response and deleted the card, nothing bad will happen now.

*5. New in editing and presenting content.*

- No matter where you switch, the editor now remembers the scroll position of all pages.

- Double-click or tap on images to open them in full-screen presentation mode. 👈

- Transitioning from editing a card to the corresponding slide in the presentation is now interconnected, just like the saved scroll position.

- New design settings ✨ to set sizes of headers, subheaders, and regular text in presentation mode.

- Preview for audio files. Support for wav, ogg, mov.

- Improved presentation interface.

And various small details and bug fixes.